Work-in-progress: Add changes handling
I use a temporary copy of the buffer held in memory to keep track of
changes.
This is because changes should be commited only after receiving the
response from the server. And since in the editor changes are naturaly
commited right before sending them to the server there was a problem
with uncommited updates that should come before the already commited
local changes.